moFontManager improved now based on FTGL, moConnectionsWindow based on wxArt2D shows Moldeo Object Patch. Lot of bugs fixed. InnoSetup doc installation updated.
git-svn-id: https://moldeo.svn.sourceforge.net/svnroot/moldeo/trunk/0.7@79 c963c68c-f84c-0410-a3d6-bef16c67260e

+++ inc/moArray.cpp
@@ -132,7 +132,7 @@ void name::Insert(int x, const T &value) {\
         arrayaux = new T [n+1];\
 \
         for( i=0,j=0 ; i < (n+1) ; i++,j++ ) {		\
-            (i==x) ? arrayaux[j--] = value : arrayaux[i] = array[j];\
+            ((int)i==x) ? arrayaux[j--] = value : arrayaux[i] = array[j];\
         }\
         arrayaux[n] = value;\
         n++;\
